+++ app/bd.php
@@ -1,5 +1,7 @@
 <?php
-require_once('config.php');
-$conn = new mysqli(SERVERNAME, USER, PASS, DB) or die("Connection failed");
-//$GLOBALS['conn'] = $conn;
+function connBD()
+{
+    $conn = new mysqli(SERVERNAME, USER, PASS, DB) or die("Connection failed");
+    return $conn;
+}
 ?>
