+++ app/controllers/login.php
@@ -25,14 +25,5 @@
             header("Location: /login");
         }
     }
-
-    function secure_input($data)
-    {
-        $data = trim($data);
-        $data = stripslashes($data);
-        $data = htmlspecialchars($data);
-        return $data;
-    }
-
     
 ?>
