+++ app/controllers/perfil.php
@@ -1,7 +1,7 @@
 <?php
 require('./models/Usuario.php');
 require('./models/Producto.php');
-require_once('./bd.php');
+$conn = connBD();
 
 isset($_SESSION['login']) ? logged($conn) : not_logged();
 $conn->close(); //Importante cerrar siempre la conexion
@@ -61,12 +61,4 @@ function mostrarProductos($idUsuario, $conn)
     }
 }
 
-function not_logged()
-{
-    ?>
-    <div class="noReg">
-    <img src="img/warning.png" alt="Atención"><p>¡Regístrate o inicia sesión para acceder!</p>
-    </div>
-    <?php
-}
 ?>
