+++ app/controllers/registrar.php
@@ -33,11 +33,4 @@ require('./img.php');
         }
     }
 
-    function secure_input($data)
-    {
-        $data = trim($data);
-        $data = stripslashes($data);
-        $data = htmlspecialchars($data);
-        return $data;
-    }
 ?>
