+++ app/index.php
@@ -1,5 +1,6 @@
 <?php
 require_once('config.php');
 require_once('bd.php');
+require_once('functions.php');
 require('routes.php');
 ?>
